requirements.txt
streamlit
boto3
langchain
openai
faiss-cpu
python-dotenv
cryptography
tiktoken
python-docx 
PyPDF2